// frontend/src/app/types/api.d.ts
export type components = {
    schemas: {
// ...
        ClientDataResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Client"];
        };
        ClientGetInput: {
            clientIds: number[];
            fields?: ("id" | "name" | "identityNumber" | "phone" | "active" | "creationDate")[];
        };
        ClientGetResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Client"][];
        };
// ...
        ClientListResponse: components["schemas"]["BaseResponse"] & {
            data?: components["schemas"]["Client"][];
            page?: number;
            total?: number;
        };
// ...
        ResultResponse: components["schemas"]["BaseResponse"] & {
            result?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: {
        ClientCreateRequest: {
            content: {
                "application/json": components["schemas"]["ClientCreateInput"];
            };
        };
        ClientDeleteRequest: {
            content: {
                "application/json": {
                    clientId: number;
                };
            };
        };
        ClientGetRequest: {
            content: {
                "application/json": components["schemas"]["ClientGetInput"];
            };
        };
        ClientListRequest: {
            content: {
                "application/json": components["schemas"]["ClientListInput"];
            };
        };
        ClientRestoreRequest: {
            content: {
                "application/json": {
                    clientId: number;
                };
            };
        };
        ClientUpdateRequest: {
            content: {
                "application/json": components["schemas"]["ClientUpdateInput"];
            };
        };
// ...
    };
    headers: never;
    pathItems: never;
};